Folder actions restructure
